pyproject.toml
[project]
name = "multiagentesanalistaposts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "canvas>=0.1.15",
    "crewai>=0.121.1",
    "crewai-tools>=0.46.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "matplotlib>=3.10.3",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "reportlab>=4.4.1",
]
